+++ NEWS.md
@@ -1,5 +1,11 @@
 # SpectriPy 0.99
 
+## Changes in 0.99.5
+
+- Add support for parameter `data` to `backendInitialize()`: this enables to
+  change the backend from a `Spectra` object to `MsBackendPy` using the
+  `setBackend()` method.
+
 ## Changes in 0.99.4
 
 - Refactor functions to initialize Python libraries.
@@ -20,7 +26,6 @@
 ## Changes in 0.99.0
 
 - Prepare for Bioconductor submission.
-
 
 # SpectriPy 0.5